Big refactoring and cleanup
