+++ src/SharpA2A.AspNetCore/A2AEndpointRouteBuilderExtensions.cs
@@ -1,6 +1,7 @@
 using DomFactory;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -14,6 +15,14 @@ public static class A2ARouteBuilderExtensions
 {
     public static readonly ActivitySource ActivitySource = new ActivitySource("A2A.Endpoint", "1.0.0");
 
+
+    /// <summary>
+    /// Enables JSONRPC A2A endpoints for the specified path.
+    /// </summary>
+    /// <param name="endpoints"></param>
+    /// <param name="taskManager"></param>
+    /// <param name="path"></param>
+    /// <returns></returns>
     public static IEndpointConventionBuilder MapA2A(this IEndpointRouteBuilder endpoints, TaskManager taskManager, string path)
     {
         var loggerFactory = endpoints.ServiceProvider.GetRequiredService<ILoggerFactory>();
@@ -21,105 +30,26 @@ public static class A2ARouteBuilderExtensions
 
         var routeGroup = endpoints.MapGroup("");
 
-        routeGroup.MapGet($"{path}/.well-known/agent.json", async context =>
+        routeGroup.MapGet($"{path}/.well-known/agent.json", (HttpRequest request) =>
         {
-            var agentUrl = $"{context.Request.Scheme}://{context.Request.Host}{context.Request.Path}";
+            var agentUrl = $"{request.Scheme}://{request.Host}{request.Path}";
             var agentCard = taskManager.OnAgentCardQuery(agentUrl);
-            await context.Response.WriteAsJsonAsync(agentCard);
+            return Results.Ok(agentCard);
         });
 
-        routeGroup.MapPost(path, requestDelegate: async context =>
-        {
-            using var activity = ActivitySource.StartActivity("HandleA2ARequest", ActivityKind.Server);
-            activity?.AddTag("endpoint.path", path);
-
-            var validationContext = new ValidationContext("1.0");
-            // Parse generic JSON-RPC request
-            var rpcRequest = await A2AProcessor.ParseJsonRpcRequestAsync(validationContext, context.Request.Body, context.RequestAborted);
-
-            // Translate Params JsonElement to a concrete type
-            IJsonRpcParams? parsedParameters = null;
-            if (rpcRequest.Params != null)
-            {
-                var incomingParams = (IJsonRpcIncomingParams)rpcRequest.Params;
-                parsedParameters = A2AMethods.ParseParameters(validationContext, rpcRequest.Method, incomingParams.Value);
-            }
-            // Ensure the request is valid
-            if (validationContext.Problems.Count > 0)
-            {
-                context.Response.StatusCode = StatusCodes.Status400BadRequest;
-                await context.Response.WriteAsJsonAsync(validationContext.Problems);
-                return;
-            }
-            // Dispatch based on return type
-            if (A2AMethods.IsStreamingMethod(rpcRequest.Method))
-            {
-                if (parsedParameters == null)
-                {
-                    context.Response.StatusCode = StatusCodes.Status400BadRequest;
-                    var errorResponse = new JsonRpcResponse()
-                    {
-                        Id = rpcRequest.Id,
-                        Error = new JsonRpcError()
-                        {
-                            Code = -32602,
-                            Message = "Invalid params"
-                        },
-                        JsonRpc = "2.0"
-                    };
-                    WriteJsonRpcResponse(context, errorResponse);
-                    return;
-                }
-                await A2AProcessor.StreamResponse(taskManager, context, rpcRequest.Id, parsedParameters);
-                await context.Response.CompleteAsync();
-            }
-            else
-            {
-                try
-                {
-                    activity?.AddTag("request.id", rpcRequest.Id);
-                    activity?.AddTag("request.method", rpcRequest.Method);
-
-                    var jsonRpcResponse = await A2AProcessor.SingleResponse(taskManager, context, rpcRequest.Id, rpcRequest.Method, parsedParameters);
-
-                    context.Response.ContentType = "application/json";
-                    if (jsonRpcResponse.Error != null)
-                    {
-                        context.Response.StatusCode = StatusCodes.Status400BadRequest;
-                        WriteJsonRpcResponse(context, jsonRpcResponse);
-                        await context.Response.CompleteAsync();
-                    }
-                    else
-                    {
-                        context.Response.StatusCode = StatusCodes.Status200OK;
-                        WriteJsonRpcResponse(context, jsonRpcResponse);
-                        await context.Response.CompleteAsync();
-                    }
-                }
-                catch (Exception e)
-                {
-                    context.Response.StatusCode = StatusCodes.Status500InternalServerError;
-                    var jsonRpcResponse = new JsonRpcResponse()
-                    {
-                        Id = rpcRequest.Id,
-                        Error = new JsonRpcError()
-                        {
-                            Code = -32603,
-                            Message = e.Message
-                        },
-                        JsonRpc = "2.0"
-                    };
-                    WriteJsonRpcResponse(context, jsonRpcResponse);
-                    await context.Response.CompleteAsync();
-
-                }
-            }
-        });
+        routeGroup.MapPost(path, ([FromBody] JsonRpcRequest rpcRequest) => A2AJsonRpcProcessor.ProcessRequest(taskManager, rpcRequest));
 
         return routeGroup;
     }
 
 
+    /// <summary>
+    /// Enables experimental HTTP A2A endpoints for the specified path.
+    /// </summary>
+    /// <param name="endpoints"></param>
+    /// <param name="taskManager"></param>
+    /// <param name="path"></param>
+    /// <returns></returns>
     public static IEndpointConventionBuilder MapHttpA2A(this IEndpointRouteBuilder endpoints, TaskManager taskManager, string path)
     {
         var loggerFactory = endpoints.ServiceProvider.GetRequiredService<ILoggerFactory>();
@@ -128,36 +58,33 @@ public static class A2ARouteBuilderExtensions
         var routeGroup = endpoints.MapGroup(path);
 
         // /card endpoint - Agent discovery
-        routeGroup.MapGet("/card", A2AHttpProcessor.GetAgentCardRequestDelegate(taskManager, logger));
+        routeGroup.MapGet("/card", async context => await A2AHttpProcessor.GetAgentCard(taskManager, logger, $"{context.Request.Scheme}://{context.Request.Host}{path}"));
 
         // /tasks/{id} endpoint
-        routeGroup.MapGet("/tasks/{id}", A2AHttpProcessor.GetTaskRequestDelegate(taskManager, logger));
+        routeGroup.MapGet("/tasks/{id}", (string id, [FromQuery] int? historyLength = 0, [FromQuery] string? metadata = null) =>
+                                            A2AHttpProcessor.GetTask(taskManager, logger, id, historyLength, metadata));
 
         // /tasks/{id}/cancel endpoint
-        routeGroup.MapPost("/tasks/{id}/cancel", A2AHttpProcessor.CancelTaskRequestDelegate(taskManager, logger));
+        routeGroup.MapPost("/tasks/{id}/cancel", (string id) => A2AHttpProcessor.CancelTask(taskManager, logger, id));
 
         // /tasks/{id}/send endpoint
-        routeGroup.MapPost("/tasks/{id}/send", A2AHttpProcessor.SendTaskMessageRequestDelegate(taskManager, logger));
+        routeGroup.MapPost("/tasks/{id}/send", (string id, [FromBody] TaskSendParams sendParams, int? historyLength, string? metadata) =>
+                                                                    A2AHttpProcessor.SendTaskMessage(taskManager, logger, id, sendParams, historyLength, metadata));
 
         // /tasks/{id}/sendSubscribe endpoint
-        routeGroup.MapPost("/tasks/{id}/sendSubscribe", A2AHttpProcessor.SendSubscribeTaskRequestDelegate(taskManager, logger));
+        routeGroup.MapPost("/tasks/{id}/sendSubscribe", (string id, [FromBody] TaskSendParams sendParams, int? historyLength, string? metadata) =>
+                                                                    A2AHttpProcessor.SendSubscribeTaskMessage(taskManager, logger, id, sendParams, historyLength, metadata));
 
         // /tasks/{id}/resubscribe endpoint
-        routeGroup.MapPost("/tasks/{id}/resubscribe", A2AHttpProcessor.ResubscribeTaskRequestDelegate(taskManager, logger));
+        routeGroup.MapPost("/tasks/{id}/resubscribe", (string id) => A2AHttpProcessor.ResubscribeTask(taskManager, logger, id));
 
         // /tasks/{id}/pushNotification endpoint - PUT
-        routeGroup.MapPut("/tasks/{id}/pushNotification", A2AHttpProcessor.ConfigurePushNotificationRequestDelegate(taskManager, logger));
+        routeGroup.MapPut("/tasks/{id}/pushNotification", (string id, [FromBody] PushNotificationConfig pushNotificationConfig) =>
+                                                                    A2AHttpProcessor.SetPushNotification(taskManager, logger, id, pushNotificationConfig));
 
-        // /tasks/{id}/pushNotification endpoint - DELETE
-        routeGroup.MapDelete("/tasks/{id}/pushNotification", A2AHttpProcessor.DeletePushNotificationRequestDelegate(taskManager, logger));
+        // /tasks/{id}/pushNotification endpoint - GET
+        routeGroup.MapGet("/tasks/{id}/pushNotification", (string id) => A2AHttpProcessor.GetPushNotification(taskManager, logger, id));
 
         return routeGroup;
-    }
-
-    internal static void WriteJsonRpcResponse(HttpContext context, JsonRpcResponse jsonRpcResponse)
-    {
-        var writer = new Utf8JsonWriter(context.Response.BodyWriter);
-        jsonRpcResponse.Write(writer);
-        writer.Flush();
     }
 }
