+++ src/SharpA2A.Core/JsonRpc/A2AMethods.cs
