+++ src/SharpA2A.Core/JsonRpc/IJsonRpcIncomingParams.cs
@@ -1,0 +1,7 @@
+using System.Text.Json;
+namespace SharpA2A.Core;
+
+public interface IJsonRpcIncomingParams : IJsonRpcParams
+{
+    JsonElement Value { get; }
+}
