+++ src/SharpA2A.Core/JsonRpc/IJsonRpcIncomingResult.cs
@@ -1,0 +1,7 @@
+using System.Text.Json;
+namespace SharpA2A.Core;
+
+public interface IJsonRpcIncomingResult : IJsonRpcResult
+{
+    JsonElement Value { get; }
+}
