+++ src/SharpA2A.Core/JsonRpc/IJsonRpcOutgoingParams.cs
@@ -1,0 +1,7 @@
+using System.Text.Json;
+namespace SharpA2A.Core;
+
+public interface IJsonRpcOutgoingParams : IJsonRpcParams
+{
+    void Write(Utf8JsonWriter writer);
+}
