+++ src/SharpA2A.Core/JsonRpc/IJsonRpcOutgoingResult.cs
@@ -1,0 +1,7 @@
+using System.Text.Json;
+namespace SharpA2A.Core;
+
+public interface IJsonRpcOutgoingResult : IJsonRpcResult
+{
+    void Write(Utf8JsonWriter writer);
+}
