+++ src/SharpA2A.Core/JsonRpc/IJsonRpcParams.cs
@@ -1,0 +1,5 @@
+namespace SharpA2A.Core;
+
+public interface IJsonRpcParams
+{
+}
