+++ src/SharpA2A.Core/JsonRpc/IJsonRpcResult.cs
@@ -1,0 +1,5 @@
+namespace SharpA2A.Core;
+
+public interface IJsonRpcResult
+{
+}
