+++ src/SharpA2A.Core/JsonRpc/InvalidParamsError.cs
@@ -1,0 +1,18 @@
+namespace SharpA2A.Core;
+
+/// <summary>
+/// Error for invalid parameters
+/// </summary>
+public class InvalidParamsError : JsonRpcError
+{
+    /// <summary>
+    /// Creates a new invalid parameters error
+    /// </summary>
+    public InvalidParamsError()
+    {
+        Code = -32602;
+        Message = "Invalid parameters";
+    }
+}
+
+
