+++ src/SharpA2A.Core/JsonRpc/InvalidRequestError.cs
@@ -1,0 +1,18 @@
+namespace SharpA2A.Core;
+
+/// <summary>
+/// Error for invalid requests
+/// </summary>
+public class InvalidRequestError : JsonRpcError
+{
+    /// <summary>
+    /// Creates a new invalid request error
+    /// </summary>
+    public InvalidRequestError()
+    {
+        Code = -32600;
+        Message = "Request payload validation error";
+    }
+}
+
+
