+++ src/SharpA2A.Core/JsonRpc/JsonRpcParams.cs
@@ -1,0 +1,16 @@
+using System.Text.Json;
+
+namespace SharpA2A.Core;
+
+internal class JsonRpcParams : IJsonRpcIncomingParams
+{
+    private readonly JsonElement jsonElement;
+
+    public JsonRpcParams(JsonElement jsonElement)
+    {
+        this.jsonElement = jsonElement;
+    }
+
+    public JsonElement Value => this.jsonElement;
+
+}
