+++ src/SharpA2A.Core/JsonRpc/JsonRpcResult.cs
@@ -1,0 +1,20 @@
+using System.Text.Json;
+
+namespace SharpA2A.Core;
+
+public class JsonRpcResult : IJsonRpcIncomingResult, IJsonRpcOutgoingResult
+{
+    private readonly JsonElement jsonElement;
+
+    public JsonElement Value => jsonElement;
+
+    public JsonRpcResult(JsonElement jsonElement)
+    {
+        this.jsonElement = jsonElement;
+    }
+
+    public void Write(Utf8JsonWriter writer)
+    {
+        writer.WriteRawValue(jsonElement.ToString());
+    }
+}
