+++ src/SharpA2A.Core/Models/AgentTypes.cs
